ui: minor: renamed BuildInstruction from BuildInsruction. Reorganized imports
