+++ components/setup/BuildInstruction.tsx
